requirements.txt
requests==2.21.0
PyJWT==1.7.1
cryptography==2.6.1